Formula evaluation has been reviewed in this commit to account for the presence of None in the data frame which translates into NULL in the database.
select * from table where variable <> 'something'

does not return rows where variable is null. Instead, the query has to be:

select * from foo where (bla <> 'something') or (bla is null)

It gets even trickier when:

select * from table where variable = 'something'

In this case the query does not need any tweaking, which you may think that in the presence of the "equals" operand, no extra clause is needed. But if the formula that contains the variable equals something is negated to NOT(variable equals something), if we do modular translation to SQL we end up with:

select * from table where NOT(variable = 'something')

and this returns exactly the same records than the first query (which was incorrect), that is:

select * from table where variable <> 'something'

The solution for this case is to insert a redundant clause in the query:

select * from table where (variable = 'something') AND (variable is not null)

That way, if the expression in the formula is negated, the new query is:

select * from table where NOT((variable = 'something' AND (variable is not null)) which returns the correct set of records.