Implement Personalized Canvas Email Action.
The action allows the creation of a plain text message. The "run"
operation requests a token, a subject, and the column with the canvas
id and sends the personalized messages to each user.

The API entry point is provided in the base.py configuration file. If
the server has more than one API entry point, it will prompt the user to
select one of them.

The integration with the API does not implement any OAuth2 protocol and
relies on the user typing a token. This is just for testing purpose
while a more comprehensive implementation of this functionality is
implemented.