Implement Personalized Canvas Email action
Implemented the functionality to edit and run the Personalized Canvas
Email action. In its "Run" step, it handles OAuth2 authentication with
the server. Stores the given token and if needed uses the refresh token
to get another one. There are still some minor changes required such as
for example a quick form to flush credentials already approved. This can
be done in the author profile page.