Add detection of users in actions with all false conditions (Issue #137)
Functionality suggested by a user that was creating an action with
a very large number of emails. In some cases, all conditions were false
which translated into an "empty" message. The platform now flags the
existence of those messages and the preview functionality offers
a link to navigate through those messages.