Implement visualization of column statistics when creating conditions
New functionality to allow users to see the column statistics when
creating a condition. Improvement suggested by users. Two options
considered, either show it as a modal in the condition tab of the
action editor, or as part of the modal creating a condition. The
second option seemed more complex because it would use a modal already
used to create the conditions. Option 1 has been selected. There is an
additional pull-down menu in the condition tab that allows the selection
of a column and opens the modal showing the stats.

Eventually this pull/down -> show stats in modal could appear in other
locations such as filter and text tab.