Allow to import a workflow taking the name from the file
This feature adds flexibility to the import operation. There could be a
case in which a workflow is imported and the user would like to keep the
original name (instead of changing it). With this enhancement the two
possibilities are available. If a name is given in the import form, it
is considered. If not, the one in the file is taken. A potential
collision between this stored name and an existing workflow is checked.