Removed requirement for redis-py-cluster
Celery backend is not properly supported when using a redis-cluster (in AWS). The underlying architecture of a redis cluster seems to require substantial changes on how Celery handles the communication. The package redis is no longer valid, and the alternative redis-py-cluster is not mature enough nor compatible with celery requirements. Removing at this point.